+++ test/test-creation.js
@@ -4,9 +4,9 @@ const path = require('path');
 const assert = require('yeoman-assert');
 const helpers = require('yeoman-test');
 const fse = require('fs-extra');
-const getFilesForOptions = require('./test-utils').getFilesForOptions;
+const getFilesForOptions = require('./utils').getFilesForOptions;
 const expectedFiles = require('./test-expected-files');
-const shouldBeV3DockerfileCompatible = require('./test-utils').shouldBeV3DockerfileCompatible;
+const shouldBeV3DockerfileCompatible = require('./utils').shouldBeV3DockerfileCompatible;
 const constants = require('../generators/generator-constants');
 const angularJsfiles = require('../generators/client/files-angularjs').files;
 const angularfiles = require('../generators/client/files-angular').files;
